# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PotholeAnalysisInput {
  /**
     * Base64-encoded image data URL. Limit to 6 MiB.
     * @maxLength 8388608
     */
  imageData: string;
}

export type PotholeAnalysisSeverity = typeof PotholeAnalysisSeverity[keyof typeof PotholeAnalysisSeverity];


export const PotholeAnalysisSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  unknown: 'unknown',
} as const;

export interface PotholeAnalysis {
  is_pothole: boolean;
  severity: PotholeAnalysisSeverity;
  /**
     * @minimum 0
     * @maximum 1
     * @nullable
     */
  confidence: number | null;
  reason: string;
  needs_manual_review: boolean;
}