# lib/api-zod/src/generated/types/potholeAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PotholeAnalysisSeverity } from './potholeAnalysisSeverity';

export interface PotholeAnalysis {
  is_pothole: boolean;
  severity: PotholeAnalysisSeverity;
  /**
     * @minimum 0
     * @maximum 1
     * @nullable
     */
  confidence: number | null;
  reason: string;
  needs_manual_review: boolean;
}