# lib/api-zod/src/generated/types/potholeAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PotholeAnalysisInput {
  /**
     * Base64-encoded image data URL. Limit to 6 MiB.
     * @maxLength 8388608
     */
  imageData: string;
}